formatting n shit

+++ chunkSaver.py
@@ -1,4 +1,4 @@
-import requests,time
+import requests, time, os, shutil
 
 def extract_data_between_quotes(lines):
     result = []
@@ -10,16 +10,16 @@ def extract_data_between_quotes(lines):
     return result
 
 def get_lines_between(lines):
-    start_index = None
-    end_index = None
+    start_index, end_index = (None,)*2
 
+    # We love hard set vars that will prob change someday
     start_str = "Promise.all(Object.keys"
     end_str = ".chunk.css\", "
 
     for i, line in enumerate(lines):
         if start_str in line and start_index is None:
             start_index = i+1
-        if end_str in line and start_index is not None:
+        elif end_str in line and start_index is not None:
             end_index = i-1
             break
     
@@ -30,10 +30,8 @@ def get_lines_between(lines):
         return []
 
 def get_the_chunks(lines:str,prefix:str):
-    #delete and create the chunks folder
-    import os,shutil
-    if os.path.exists('chunks'):
-        shutil.rmtree('chunks')
+    # Delete and create the chunks folder just incase he changes chunks to be lower
+    if os.path.exists('chunks'): shutil.rmtree('chunks')
     os.makedirs('chunks', exist_ok=True)
 
     lines = lines.splitlines()
@@ -42,7 +40,7 @@ def get_the_chunks(lines:str,prefix:str):
 
     for i, line in enumerate(data_between_quotes):
         print(f"{i+1}: {prefix}.{i+1}.{line}.chunk.js")
-        time.sleep(0.1)
+        time.sleep(0.1) # Just to not send them super fast (prob doesn't even matter tbh)
         request = requests.get(f'https://bloxd.io/static/js/{prefix}.{i+1}.{line}.chunk.js')
         with open(f'chunks/{i+1}.chunk.js', 'wb') as f:
             f.write(request.content)
